loader/requirements.txt
azure-storage-blob
sqlalchemy
pyodbc
python-dotenv